moved around the spell list to make more sense
