+++ src/gameplay/commandServices/PlayerService.java
@@ -2,6 +2,7 @@ package gameplay.commandServices;
 
 import gameplay.StatModMethods.PlayerStatMethods;
 import gameplay.commandServices.utilities.TakeUtilities;
+import pojos.Ability;
 import pojos.items.enums.ArmorMaterial;
 import pojos.items.enums.WeaponType;
 import uiView.UIMain;
@@ -295,6 +296,22 @@ public class PlayerService {
 		return output.toString();
 	}
 
+	public String listAbilities() {
+		StringBuilder outputBuilder = new StringBuilder();
+		outputBuilder.append(UIMain.player.getName() + " Abilities, Skills, and Spells");
+		outputBuilder.append("\n\n");
+		for(Ability ability : UIMain.player.getSkills()) {
+			outputBuilder.append(String.format("%-5s",  "NAME: " + ability.getName()));
+			outputBuilder.append(String.format("%20s", "ATTRIBUTE: " + ability.getAttribute()));
+			if(ability.getHpCost() > 0) {
+				outputBuilder.append(String.format("%30s", "HP: " + ability.getHpCost()));
+			}
+			if(ability.getSpCost() > 0) {
+				outputBuilder.append(String.format("%30s", "SP: " + ability.getSpCost()));
+			}
+		}
+		return outputBuilder.toString();
+	}
 	private String takeItemFromCell(String param) {
 		StringBuilder outputBuilder = new StringBuilder();
 		if (param == null || param.equals("")) {
